requirements.txt
requests~=2.32.0
google-generativeai
openai~=1.61.1